requirements.txt
numpy>=1.24.0
Pillow>=9.0.0
groq>=0.4.0
python-dotenv>=1.0.0
assemblyai>=0.30.0

requests>=2.28.0
